Refactor: 현재위치값 헤더로 이동
